requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.3
future==0.18.2
gunicorn==20.0.4
idna==2.8
line-bot-sdk==1.15.0
pathspec==0.7.0
psycopg2==2.8.4
pytz==2019.3
regex==2019.12.20
requests==2.22.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
dj_static==0.0.6
dj_database_url==0.5.0
flask
